requirements.txt
Django
django-extensions
werkzeug
sorl-thumbnail
Pillow
gitpython
django-widget-tweaks
python_slugify==1.2.4
django-simple-captcha
